[MISC] move test pipeable_config_element_test.cpp
