+++ test/unit/alignment/configuration/align_config_method_test.cpp
@@ -12,7 +12,7 @@
 #include <seqan3/alignment/configuration/align_config_method.hpp>
 #include <seqan3/core/configuration/configuration.hpp>
 
-#include "../../core/algorithm/pipeable_config_element_test_template.hpp"
+#include "../../core/configuration/pipeable_config_element_test_template.hpp"
 
 // ---------------------------------------------------------------------------------------------------------------------
 // pipeable_config_element_test template
