+++ test/unit/alignment/configuration/align_config_on_result_test.cpp
@@ -12,7 +12,7 @@
 
 #include <seqan3/alignment/configuration/align_config_on_result.hpp>
 
-#include "../../core/algorithm/pipeable_config_element_test_template.hpp"
+#include "../../core/configuration/pipeable_config_element_test_template.hpp"
 
 // -----------------------------------------------------------------------------
 // test template : pipeable_config_element_test
