+++ test/unit/alignment/configuration/align_config_parallel_test.cpp
@@ -13,7 +13,7 @@
 
 #include <seqan3/alignment/configuration/align_config_parallel.hpp>
 
-#include "../../core/algorithm/pipeable_config_element_test_template.hpp"
+#include "../../core/configuration/pipeable_config_element_test_template.hpp"
 
 // ---------------------------------------------------------------------------------------------------------------------
 // test template : pipeable_config_element_test
