+++ test/unit/alignment/configuration/align_config_score_type_test.cpp
@@ -14,7 +14,7 @@
 #include <seqan3/test/expect_same_type.hpp>
 #include <seqan3/utility/type_traits/basic.hpp>
 
-#include "../../core/algorithm/pipeable_config_element_test_template.hpp"
+#include "../../core/configuration/pipeable_config_element_test_template.hpp"
 
 template <typename test_t>
 struct align_cfg_output_test : public ::testing::Test
