+++ test/unit/core/algorithm/CMakeLists.txt
@@ -1,4 +1,3 @@
 seqan3_test (algorithm_result_generator_range_test.cpp)
-seqan3_test (pipeable_config_element_test.cpp CYCLIC_DEPENDING_INCLUDES include-seqan3-core-configuration-configuration.hpp)
 
 add_subdirectories()
