+++ test/unit/core/configuration/CMakeLists.txt
@@ -1,3 +1,2 @@
 seqan3_test (configuration_test.cpp)
-
-add_subdirectories()
+seqan3_test (pipeable_config_element_test.cpp CYCLIC_DEPENDING_INCLUDES include-seqan3-core-configuration-configuration.hpp)
