+++ test/unit/search/configuration/on_result_test.cpp
@@ -9,7 +9,7 @@
 
 #include <seqan3/search/configuration/on_result.hpp>
 
-#include "../../core/algorithm/pipeable_config_element_test_template.hpp"
+#include "../../core/configuration/pipeable_config_element_test_template.hpp"
 
 // ---------------------------------------------------------------------------------------------------------------------
 // test template : pipeable_config_element_test
